# server/docs/swagger.yaml
definitions:
...
  models.Folder:
    properties:
      _id:
        type: string
      color:
        type: string
      eventIds:
        items:
          type: string
        type: array
      isDeleted:
        type: boolean
      name:
        type: string
      userId:
        type: string
    type: object
...
  routes.CreateFolderResponse:
    properties:
      id:
        type: string
    type: object
host: localhost:3002/api
info:
  contact: {}
  description: This is the API for Schej.it!
  license: {}
  title: Schej.it API
  version: "1.0"
paths:
...
  /user/folders:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: A list of all folders for the user
          schema:
            items:
              $ref: '#/definitions/models.Folder'
            type: array
        "400":
          description: Invalid user ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to get folders
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all folders
      tags:
      - folders
    post:
      consumes:
      - application/json
      parameters:
      - description: Folder name and optional color
        in: body
        name: payload
        required: true
        schema:
          allOf:
          - type: object
          - properties:
              color:
                type: string
              name:
                type: string
            type: object
      produces:
      - application/json
      responses:
        "201":
          description: The ID of the created folder
          schema:
            $ref: '#/definitions/routes.CreateFolderResponse'
        "400":
          description: Invalid user ID or request body
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to create folder
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new folder
      tags:
      - folders
  /user/folders/{folderId}:
    delete:
      parameters:
      - description: Folder ID
        in: path
        name: folderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200": {}
        "400":
          description: Invalid user ID or folder ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to delete folder
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete a folder
      tags:
      - folders
    get:
      parameters:
      - description: Folder ID
        in: path
        name: folderId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: The folder object with events
          schema:
            $ref: '#/definitions/models.Folder'
        "400":
          description: Invalid user ID or folder ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Folder not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to get events in folder
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a folder by its ID and its contents
      tags:
      - folders
    patch:
      consumes:
      - application/json
      parameters:
      - description: Folder ID
        in: path
        name: folderId
        required: true
        type: string
      - description: New folder name and/or color
        in: body
        name: payload
        required: true
        schema:
          allOf:
          - type: object
          - properties:
              color:
                type: string
              name:
                type: string
            type: object
      produces:
      - application/json
      responses:
        "200": {}
        "400":
          description: Invalid user ID or folder ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to update folder
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update a folder's name or color
      tags:
      - folders
...
swagger: "2.0"